requirements.txt
langgraph>=0.4.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-tavily>=0.2.0
python-dotenv>=1.0.0
rich>=14.0.0
pydantic>=2.0.0
